Now serial uses a tree data structure instead of a zipper.
The downside is that now it no longer has constant access time for small
steps, but the upside is that the algorithm is much simpler and shorter
and going backwards is no longer a linear function of the number of steps
back being taken, and furthermore logarithmic time isn't so bad.  It still
takes linear time to travel forward, though, in order to ensure that the
future animation has the correct starting state.