Reworked AnimationM.
I pulled the implementation into an internal module that is now selectively
exported by the new Slick.AnimationM module.  I also simplified a lot of
things when I realized that a lot of functions could be consolidated into
a single appendAnimation function, which is smart enough to do all of the
right things.